feature: first rudimentary version of get_json()
The function get_json() is going to be our workhorse, doing most of the work required to turn an email message into an Elasicsearch document; this is the first version, and it just grabs the sender and body and builds the basic document. We will need to figure out the other attributes, etc.

...