feature: test code to load and use the LSI model
We wrote this code in isolation to test the LSI model we built and saved separately. We did this to test a fix for the __getitem__, which was working for LDA but not LSI.
It turned out the solution was to use the LsiModel.load() instead of the generic SaveLoad.load().

...