feature: added support for a sender exclusion list
If we know that a particular sender is a newsletter or an advertisement we suspect it doesn't contribute to any meaningful cluster.
Unfortunately in order to do this we need to introduce use of pyzmail, and it needs the original raw email data.
When we do this it looks like we've reintroduced a bunch of HTML noise that will need to be ironed out.
So this is still a work in progress.

...