feature: buildout of several features
We're no longer using print() for anything important; everything uses the logger.
The methods now run in a loop and write their results to separate output files in the output folder
Finally, I made an attempt to set aside members for each model.

Unfortunately the logentropy model needs too much memory to be usable for the number of documents we have.
And finally, this example clusters topics, not documents, so it isn't what we need.

...