feature: tfidf min_df and max_df are now settings
We've also added minimal support for multi-dimensional scaling and Ward multilevel clustering,
but the multilevel clustering is running into memory issues so it is currently inactive.
Finally, we're using the accessor for the specific stopwords to refine our bag of words.
This is not a great solution; it is prone to overfitting.

...