feature: added a tokenizer that uses noun phrases
Added a tokenizer that extracts the noun phrases from a document and then the tokens from those phrases
Changed the calling code to use it.
Added some more logging to help us track where the time is going.
Not surprisingly we're spending almost all our time training the tfidf model, because that's where we're calling the tokenizer,
and using the parts of speech tagger is really expensive.

Finally, changed the plot to use a dot instead of an X to reduce chartjunk.

https://github.com/mikedelong/aarhus/issues/4